tests/wpt/web-platform-tests/html/canvas/element/path-objects/2d.path.arcTo.shape.curve1.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.path.arcTo.shape.curve1</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.arcTo.shape.curve1</h1>
<p class="desc">arcTo() curves in the right kind of shape</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("arcTo() curves in the right kind of shape");
_addTest(function(canvas, ctx) {

var tol = 1.5; // tolerance to avoid antialiasing artifacts

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);

ctx.strokeStyle = '#f00';
ctx.lineWidth = 10;
ctx.beginPath();
ctx.moveTo(10, 25);
ctx.arcTo(75, 25, 75, 60, 20);
ctx.stroke();

ctx.fillStyle = '#0f0';
ctx.beginPath();
ctx.rect(10, 20, 45, 10);
ctx.moveTo(80, 45);
ctx.arc(55, 45, 25+tol, 0, -Math.PI/2, true);
ctx.arc(55, 45, 15-tol, -Math.PI/2, 0, false);
ctx.fill();

_assertPixel(canvas, 50,25, 0,255,0,255, "50,25", "0,255,0,255");
_assertPixel(canvas, 55,19, 0,255,0,255, "55,19", "0,255,0,255");
_assertPixel(canvas, 55,20, 0,255,0,255, "55,20", "0,255,0,255");
_assertPixel(canvas, 55,21, 0,255,0,255, "55,21", "0,255,0,255");
_assertPixel(canvas, 64,22, 0,255,0,255, "64,22", "0,255,0,255");
_assertPixel(canvas, 65,21, 0,255,0,255, "65,21", "0,255,0,255");
_assertPixel(canvas, 72,28, 0,255,0,255, "72,28", "0,255,0,255");
_assertPixel(canvas, 73,27, 0,255,0,255, "73,27", "0,255,0,255");
_assertPixel(canvas, 78,36, 0,255,0,255, "78,36", "0,255,0,255");
_assertPixel(canvas, 79,35, 0,255,0,255, "79,35", "0,255,0,255");
_assertPixel(canvas, 80,44, 0,255,0,255, "80,44", "0,255,0,255");
_assertPixel(canvas, 80,45, 0,255,0,255, "80,45", "0,255,0,255");
_assertPixel(canvas, 80,46, 0,255,0,255, "80,46", "0,255,0,255");
_assertPixel(canvas, 65,45, 0,255,0,255, "65,45", "0,255,0,255");


});
</script>
